+++ examples/connect-app-teleop-react/index.html
@@ -4,10 +4,10 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Viam App and Tele-operation Demo</title>
-    <link rel="stylesheet" href="/frontend/index.css" />
+    <link rel="stylesheet" href="/src/index.css" />
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/frontend/main.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
