+++ examples/connect-app-teleop-react/src/components/locations-list.tsx
@@ -17,7 +17,7 @@ export const LocationsList = (props: LocationsListProps): JSX.Element => {
       setLocations(locations);
     }
     getLocations();
-  }, [appClient]);
+  }, [appClient, orgId]);
 
   return (
     <div>
