+++ examples/connect-app-teleop-react/src/components/machine-part-control.tsx
