+++ examples/connect-app-teleop-react/src/components/machine-parts-list.tsx
@@ -17,7 +17,7 @@ export const MachinePartsList = (props: MachinesPartsListProps): JSX.Element => 
       setMachineParts(machineParts);
     }
     getMachineParts();
-  }, [appClient]);
+  }, [appClient, machine.id]);
 
   return (
     <div>
