+++ examples/connect-app-teleop-react/src/components/machines-list.tsx
