+++ examples/connect-app-teleop-react/src/components/video-stream.tsx
