+++ examples/connect-app-teleop-react/src/vite-env.d.ts
